Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
python-dotenv==1.0.0
marshmallow==3.21.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.7
requests==2.31.0
Werkzeug==2.3.7
python-decouple==3.8
Pillow==10.1.0
bcrypt==4.1.2